+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 ## Changelog
 
+`react-js-cron` follows [Semantic Versioning 2.0.0](http://semver.org/).
+
+#### 5.0.1
+
+- Fix `dropdownsConfig` `filterOption` prop not working for months and week-days.
+
+#### 5.0.0
+
+- **(Breaking change!)** Fix issue [#60](https://github.com/xrutayisire/react-js-cron/issues/60): react-js-cron now only support antd >= v5.8.0, use "suffixIcon" instead of deprecated "showArrow" for Select component
+
+#### 4.1.0
+
+- **(New feature)** Issue [#36](https://github.com/xrutayisire/react-js-cron/issues/36): Dropdowns specific configuration (mode, allowClear, filterOption, ...)
+- Fix dropdowns popover style
+
+#### 4.0.0
+
+- **(Breaking change!)** Fix issue [#55](https://github.com/xrutayisire/react-js-cron/issues/55): react-js-cron now only support antd >= v5.5.0, change antd Select property "dropdownMatchSelectWidth" to "popupMatchSelectWidth"
+
 #### 3.2.0
 
 - Add themes support for antd v5 with ConfigProvider (@nefedov-dm)